some sonarqube fixing

+++ src/main/java/server/Server.java
@@ -216,7 +216,7 @@ public class Server
 			if(timer==null){
 				timer=new Timer();
 				System.out.println("START TIMER");
-				timer.schedule(new StartingGameTimerTask(this,oneRoomLobby, playersView),20*1000);
+				timer.schedule(new StartingGameTimerTask(this,oneRoomLobby, playersView),(long)(20*1000));
 			}
 		}	
 	}
